+++ includes/slm-api-listener.php
@@ -15,6 +15,10 @@ class SLM_API_Listener {
     function __construct() {
         if (isset($_REQUEST['slm_action']) && isset($_REQUEST['secret_key'])) {
             //This is an API query for the license manager. Handle the query.
+
+            //Trigger an action hook
+            do_action('slm_api_listener_init');
+
             $this->creation_api_listener();
             $this->activation_api_listener();
             $this->deactivation_api_listener();